docs: reorganize repo for public tutorial presentation
